Firefly-1341: enhance embedded position search panel for upload
