+++ src/firefly/js/templates/hydra/HydraManager.js
@@ -6,14 +6,16 @@ import {take, fork} from 'redux-saga/effects';
 import {SHOW_DROPDOWN, SET_LAYOUT_MODE, getLayouInfo,
         dispatchSetLayoutInfo, dropDownManager} from '../../core/LayoutCntlr.js';
 import {removeChartsInGroup} from '../../charts/ChartsCntlr.js';
-import {TABLE_SEARCH, TBL_RESULTS_ADDED, TABLE_REMOVE} from '../../tables/TablesCntlr.js';
+import {TABLE_SEARCH, TBL_RESULTS_ADDED, TABLE_REMOVE, TABLE_LOADED, TBL_RESULTS_ACTIVE
+} from '../../tables/TablesCntlr.js';
 import {visRoot, dispatchDeletePlotView} from '../../visualize/ImagePlotCntlr.js';
-import {removeTablesFromGroup, getAllTableGroupIds, smartMerge} from '../../tables/TableUtil.js';
+import {removeTablesFromGroup, getAllTableGroupIds, smartMerge, getTblById} from '../../tables/TableUtil.js';
 import {getSearchInfo} from '../../core/AppDataCntlr.js';
 import ImagePlotCntlr from '../../visualize/ImagePlotCntlr.js';
 import {CHART_ADD} from '../../charts/ChartsCntlr.js';
 import {REPLACE_VIEWER_ITEMS} from '../../visualize/MultiViewCntlr.js';
 import {deleteAllDrawLayers} from '../../visualize/DrawLayerCntlr';
+import {MetaConst} from 'firefly/data/MetaConst';
 
 /**
  * Configurable part of this template
@@ -54,6 +56,13 @@ export function* hydraManager() {
         newLayoutInfo = smartMerge(layoutInfo, {showImages, showXyPlots, showTables});
 
         switch (action.type) {
+            case TBL_RESULTS_ADDED:
+            case TABLE_LOADED:
+            case TBL_RESULTS_ACTIVE:
+                const tbl = getTblById(action.payload.tbl_id);
+                if (tbl?.request?.META_INFO?.[MetaConst.UPLOAD_TABLE]) break; //to not change layoutInfo if it is an upload table
+                //intentional fallthrough
+            case CHART_ADD:
             case TABLE_SEARCH:
                 newLayoutInfo = handleNewSearch(newLayoutInfo, action);
                 break;
