+++ src/firefly/js/visualize/ui/CommonUIKeys.js
@@ -1,3 +1,7 @@
 export const CONE_CHOICE_KEY = 'CONE';
 export const POLY_CHOICE_KEY = 'POLY';
+export const UPLOAD_CHOICE_KEY = 'UPLOAD';
 export const CONE_AREA_OPTIONS = [{label: 'Cone', value: CONE_CHOICE_KEY}, {label: 'Polygon', value: POLY_CHOICE_KEY}];
+
+export const CONE_AREA_OPTIONS_UPLOAD = [{label: 'Cone', value: CONE_CHOICE_KEY}, {label: 'Polygon', value: POLY_CHOICE_KEY},
+    {label: 'Multi-object', value: UPLOAD_CHOICE_KEY}];
